Move Exporter to its own module
