+++ libs/telemetry/src/capturing/ng.rs
@@ -1,6 +1,7 @@
 pub mod collector;
+pub mod exporter;
 pub mod layer;
 pub mod traceparent;
 
-pub use collector::Exporter;
+pub use exporter::Exporter;
 pub use layer::layer;
